Add FlowUnitsConverter and ability to relax invariants check
Invariants checks are important in ensuring that we're not creating or destroying energy (or other conservation value) through a component. However, for unit conversion, the values of the magnitudes should not match (necessarily) across the unit conversion component. This allows us to use the FlowElement mechanism and FlowState while disabling invariants checks under special circumstances.