Add explicit default constructor to get around C++ compiler issue
As best I can understand, something is going on with StreamTypes used as Elements in containers. By creating constructors that require arguments (but not specifying the 0 argument constructor), C++ (rightly) no longer creates the 0-argument constructor implicitly. There must be some code in containers that depends on creating a "blank" object and using a move/copy constructor to initialize it. We assume the implicit copy constructor is fine in this case. Note: I'm not super happy with allowing a default "electricity" streamtype as this could cause some wierdness in the code; it's a bit of a compromise to allow forward movement until we understand a better approach...