Ability to read input file and translate components to basic elements
For each component id, creates a list of elements that comprise that component. We are currently using a vector but a class or unordered_map could also be used in the future. The essential idea is that the input file will have enough information to tease out what elements comprise a component and those elements have to be interconnected (eventually) in specific ways (i.e., a template).