Update for build on Windows
Remove /WX flag on Windows as this causes warnings to be treated as errors and it is difficult to resolve all warnings on the window side. Specifically, there are some warnings from the dependencies that are getting flagged that are difficult to do something about. Otherwise, main changes are to add if constexpr and wrap all blocks with "{...}" even if it is a single statement (for consisent style) and also address size_type issue in test code. Also prevent, "unused parameter" warnings.