Add TODO item as reminder for possible way to handle distributions in the future.
This is a reminder that when distributions are created, we don't have all the information we would like as of yet. In particular, it would be good for a single random number generator to be created in Main (this could be explicitly seeded or seeded with the clock). Thus, when TOML is being read, we don't really want to have to create a random number generator. This jots down some ideas of how to handle that as a TODO item. Note: we may need to utilize unique_pointers and clone.