Inprocess checkin prior to refactoring.
All unit tests are passing but that is due to disabling pieces in TestFragilityWorksForNetworkSim. We know that simulating with more than one Main instance doesn't work but it should. Getting ready to do some refactoring and so I want a check-in to verify we don't loose any ground and tests still pass after refactoring.