Fix Mux and FlowMeter issues with event propagation.
FlowMeter was not handling multiple requests at the same time well. Although that should never happen, FlowMeter needs to be robust to the condition of multiple requests in practice. Similarly, for Mux, we needed to ensure that communication occurs whenever one of the input or output ports is updated. We also added a tracking array for prev_outputs and prev_inputs to allow us to detect when new information needs to be propagated.