Update FlowElement and Muxer model to pass test.
This new test exercises a case where an element requests an upstream value and is at first denied but, because conditions change elsewhere in the network, is later able to be satisifed. Elements have to keep track of what they are ultimately requesting to ensure invariants hold. You can always achieve an inflow up to the requested amount either instantly or at a later time.