Update example 3 due to TOML constraints.
The C++ library for parsing TOML doesn't allow for variable types in a single array (even though the TOML language specification does). Therefore, we are quoting everything in the network definition array. Also, the specification for the array hookup for example 3 was incomplete and incorrect per the specification; that has been corrected.