Update tests and examples to output units.
Switch to be rigorous about units. Inside the tool, we use seconds, kW, and kJ. Input is read according to the units specified in the input files. Units are writte according to fixed units for the time being although that should be user specifiable (i.e., we should read them out of the simulation_info block of the input file -- just not there yet!).