Fix sink functionality to specify 0 load at end of load profile.
The Sink object (erin_element.cpp) was not broadcasting upstream when it's load went to zero at end of load profile. This corrects that issue and also ensures that results include both the 0-time (i.e., start) and max_time entries. Updated tests to correspond as well as regression test files.