Update ex 3 c4 hurricane to be deterministic; add regression test.
Ran and reviewed the output of the example 3 class 4 hurricane run. Here, the intensities have been set so high as to guarantee failure of all components due to fragility curve failure (0% chance of survival). This case essentially tests in a deterministic way that all components simulate correctly in their "failed" state.