Update days per year to correspond to a non-leap-year model.
This is perhaps contentious but we are assuming 365 days per year (not, 365.25 days per year). That means that "1000 years" will not quite be an accurate amount of time but it does mean that we'll have a clean and simple way to account for time with elapsed days and months. We may change this in the future...