Add stubs for getting data from AllResults to test.
The direction we're going here is to introduce a "total energy availability" which is the ratio of total energy achieved divided by total energy requested for each stream over all loads (i.e., individual components are dropped out). We'll use this as a metric (along with some others) to confirm that we're able to control (i.e., mock) the results of random processes in the input.