Make time_unit specification consistent.
We had several regression test files that were working "by accident" in that they specified a time_unit value that happened to be the default but which was not actually read (time_units = "..." was written but only time_unit = "..." is read). Make time_unit declaration in the occurrence_distribution consistent with everything else.