Add ex06 and add regression tests.
Example 6 uncovered some issues with elements that request flow from upstream, are initially denied, and then later are able to be satisfied. In a couple of cases, later supplying that flow caused a false exception in mux (because mux was looking at previous inflows vs the total outflow request) and issues with when lossflow is reported and in the implementation of the mux "in-order" dispatch (which needs to go off of outflow request, NOT the previous outflow values). These have all been rectified including setting the precision higher on CSV output.