Major refactor to switch to port-level logging.
This refactor was a bit bigger and more "all-at-once" that I would have liked in hindsight. A summary of the changes is as follows: all elements now accept port-level logging which is enabled (via set_recording_on()) when components are transformed to elements; the way the output flows are pulled back has been updated as well (more work is done in the FlowWriter which has all the information); FlowWriter has been updated to track component type of ports that register themselves; network functions have been updated to take an optional two_way boolean for connecting elements both ways; reworked add_to_network of components to elements, update refactorings.