Fix issue with find_or / toml::table going into iteration even when size is 0
This is bizarre and only happens on Windows (not Mac OS X / clang). Despite a toml::table having size 0, it still enters the for-loop and provides a pair that points to bad memory. I thought that toml::table was just an alias for an unordered_map so I find this extremely bizarre and disconcerting but, at any rate, this seems to fix the issue.