Fix id error.
Here, we should have been storing the subtype id, NOT the cdf id. The subtype id is the index into fixed_cdf whereas the id is the index into cdfs. The index into cdfs needs to be returned but the index into fixed_cdf needs to be stored so that it can be looked up later when cdf_id is used.