Correct uncontrolled source behavior.
The uncontrolled source was jumping ahead too quickly and getting out of time sequence with other blocks. Now, it only reports out when an external event happens OR, when a reportable internal transition, it schedules a dt=0 step event to report out on. This slowing down allows other blocks to get to the same time sequence. This area of timing in devs needs to be dug into further.