Change load profile to apply to component, not location; add ids to all component tables.
With the (soon) introduction of uncontrolled sources, a load profile can no longer be unambiguously applied with only a location and flow, a component id is needed. As such, this changes the definition of the load-profile csv table and processing to specify and work off of a component_id (vs a "building_id"). We furthermore also add explicit ids to the component tables.