Change how LoadItem is implemented
Drop the dual nature of LoadItem to both represent the load end AND the load itself. Now, the interetation of LoadItem is that the load at the specified time IS the specified value and it remains at that value until another event occurs. There is no longer a need to have an ending LoadItem (li.get_is_end(), etc.).