Add fix.
This removes the if (diff > tol) clause and further restricts when ports are set as unchanged by requiring the time of last change to be the current time. This does appear to fix the issue at the increase of some chattiness. Need to further enhance the unit test for mux to ensure this is working correctly.