Rename failure_cdf and repair_cdf to failure_dist and repair_dist
Again, favor "dist" (distribution) over the specific term "CDF" (cumulative distribution function). We are not actually using the CDF in our project. Instead, we typically use the "Quantile" for internal calculations but users can input the mathematical parameters for the distributions they desire and that's sufficient. So it's better to think of things as distributions.