Update tests and example 4 output
Remove unused variable and throw for
failure to provide TOML table.
Combine all tests into one ctest package.
Don't expect ids to be the same (tags should
be the same but no guarantee on ids).
Update the output for example 4. Example
4 is subject to loading order-of-operations
in terms of the sampled "random numbers"
from the provided series. Any time we
update the loading operations, it may
change. Note, hover, that the statistical
energy availability values are roughly the
same between different checked-in runs.