Rework available flows for converter non-outflows
This is specifically for the lossflow and wasteflow. Lossflow
is flow that can be used for other purposes while wasteflow
is flow that goes to waste. The "available" flow for these
(i.e., non-outflow), must equal the difference between the
actual inflow and actual outflow.

The rule for available flow is that if flow is available, a
downstream component should be able to switch to it without
the available flow stream being recalculated. For a converter,
what's available as loss and wasteflow is the difference between
the actual inflow and outflows...