Add new example and correct efficiency
Example 5 is a ladder of multiple loads and converters. It showcases how a low efficiency
converter can be chained to yield a higher efficiency.

We also correct an issue in efficiency reporting; efficiency should be based on the
requested output (not the achieved). However, we should use the achieved values for
checking that we sum to zero. We also report the "delivery effectiveness" which is the
actual delivery to outflows over the requested as a percentage.