Refactor code to pass test 9
Added another pass so now we have a backward pass (load to source), a forward pass
(source to load), and a final pass after finalization (this is after the actual flows have been
determined). This allows the code in each pass to become drastically simpler.

Also fixed a bug with setting initial conditions with a Store as a Store acts like both a
source (discharge) and a load (charge request) and so also added those fields to the flows
upon initialization.