Add stress test for erin next and fix performance issue
The stress test defaults to testing 5,000 source/load pairs with the load being a
schedule-based load. The schedule is 8,760 hours. We check the simulation time
for conducting the simulation.

The schedule-based load object was updated so that we store the index into the
TimeAndLoad array structure so that we don't have to do loop iteration to find the
next schedule-based load event.