Added finalTime to simulation, Fixed Mux request strategy, removed debug counting; refactored code.
This commit mainly is focused on getting the Mux component to simulate correctly.
We moved in the algorithm from ERIN 0.55 for apportioning mux requests to inports.
We have also split out the lossport and wasteport functions.

NOTE: we are currently allowing the converter to have efficiency > 1 to accommodate chillers and heat
pumps and other "COP" based components. However, I believe we should add a separate component
for these items as a "COP" component should not have lossport/wasteport.