Hook random processes into simulation and TOML parsing
We can now set the random process by parsing the simulation_info structure in TOML.
We also update the examples to NOT specify network=".." under scenarios as that is no
longer needed in a single-network configuration.

We also correct an issue with year consistency. One year is considered 365 days (not 365.25) in
ERIN. There was an inconsistency in time calculations due to using different conversion factors.
That has been corrected.