Add an ignore construct to explicitly ignore unsued variable instances.
The assert mechanism in C++ automatically compiles out to a no-op
under release mode. However, if we have variables that are only
created to check assertions, they are flagged as compiler warnings
under release. Using a 2009 blog post from Herb Sutter (see reference
below), we created a templated ignore construct. This allows us
to explicitly ignore and communicate a variable is being ignored.

The only valid known use case for this is when a variable is
created for an assert.

Reference: https://herbsutter.com/2009/10/18/mailbag-shutting-up-compiler-warnings