Enable time to be like "Groundhog's Day"
ERIN simulations simulate the same year (or,
more precisely, scenario periods) over and
over again. Thefore, a rare occurrence that
happens "90 years from now" is actually happening
on the 90th occurrence of the current year.

This fixes the time system so that reliability
curves for components are regenerated for each
scenario occurrence.