Update ApplyReliabilitiesAndFragilities to calc directly
The s.TheModel.Reliabilities (std::vector<ScheduleBasedReliability>)
are now directly calculated. We also caught a place where the
random function was NOT being used to calculate the repair dist
time for a fragility failure and that has been corrected.

This version should be easier to test as an entire Simulation
object is not required.