Add Courier and logging system.
This adds a default Courier instance and shows how
to hook it up with this library. Note: we are currently
overriding the error behavior of Courier as we don't
currently want to exit when an error is logged as we might have
additional clean-up or activities prior to exit.

This also includes some small clean-up / linting fixes
suggested by CLion as well as light testing on the logging
system. Those tests have been wired into regress.py.