Refactor common Courier gen; more logging; fix bug with GCC w/ LogLevel.
GCC was apparently mistaking having a field with the same name
as a type as a redefinition of that type. By renaming the field,
we should be able to get around that.

Logging was passed into the TOML parsing routines. More work
is still needed there but this gives us a good start.