Add reporting to each component.
If a connection's from or to component is report=true,
then that connection will write in event file. Otherwise,
if both from and to component are report=false, it will
not write. Component specific events (SOC, storage amount)
will also not write if report = false.

This commit also did some light investigation into intersection
with groups but more testing is needed.