Update to do energy correction due to floating point rounding issues.
The dt * power flow is causing a "wrong-rounding" that pushes the energy use slightly over or under the clamp targets for availableCharge and availableDischarge. This change just clamps to those maximum values and issues a warning. The comprehensive fix is probably to switch to 100% integer math to ensure calculations stay correct across equations.