Track load requested and source input energy.
This adds critical missing information to the
statistics file that informs the end user of
the total source energy input and the requested
load energy output. This is needed to better
understand the energy flows (for cost modeling)
and to pair with load not served to understand
what was originally requested.